+++ scripts/update_pavlov_stats_db.bat
